Move primitive conversions to conversion module
Moved conversions:
- Metadata from primitive
- Primitive from native type
