+++ xdl_rs/src/xdl_primitive/primitive_test.rs
@@ -198,26 +198,3 @@ fn string_deserialize_works() {
     let primitive = XdlType::deserialize_type(&metadata, &mut reader).unwrap();
     assert_eq!(primitive, expected);
 }
-
-#[test]
-fn test_xdl_primitive_from_bool() {
-    let value: bool = true;
-    let primitive: XdlPrimitive = value.into();
-    assert_eq!(primitive, XdlPrimitive::Bool(true));
-}
-
-#[test]
-fn test_xdl_primitive_from_u8() {
-    let value: u8 = 42;
-    let primitive: XdlPrimitive = value.into();
-    assert_eq!(primitive, XdlPrimitive::U8(42));
-}
-
-// Add similar tests for other primitive types...
-
-#[test]
-fn test_xdl_primitive_from_string() {
-    let value: String = "Hello, World!".to_string();
-    let primitive: XdlPrimitive = value.into();
-    assert_eq!(primitive, XdlPrimitive::String("Hello, World!".to_string()));
-}
